pyproject.toml
[project]
name = "rita-mcp-server"
version = "0.2.0"
description = "RITA: Nifty 50 Double DQN Investment System — MCP + Python Client"
authors = [
    {name = "RITA Team"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0,<2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ta>=0.11.0",
    "stable-baselines3>=2.3.0",
    "gymnasium>=0.29.0",
    "torch>=2.0.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.0.0",
    "streamlit>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rita"]

[project.scripts]
rita = "rita.interfaces.mcp_server:main"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
